ci(restack): green clippy/typos/tests/file-size on #203
- clippy: drop unused import (bf16 test), #[allow(too_many_arguments)] on
  copy_d2d_2d_async, #[allow(needless_range_loop)] + drop redundant f64 cast
  in the flashinfer CPU-reference test, add varlen (main #186 param) to the
  14 check_kernel_batched_eligible test calls
- typos: allow CUTLASS float_ue4m3_t / ue4m3 identifiers
- test: buffers alloc_count 27 -> 29 (Holo fp8_act + fp8_act_scale)
- file-size: re-add prefill_a_step.rs allow-entry; factor model_a.rs scalar
  dequant into read_scalar_weight_scale helper (508 -> 487 LoC)

+++ crates/spark-runtime/src/flashinfer.rs
@@ -283,6 +283,7 @@ mod tests {
 
     #[test]
     #[ignore = "requires a free CUDA device + FLASHINFER_HOME build"]
+    #[allow(clippy::needless_range_loop)] // numerical reference: index-parallel loops read clearest
     fn flashinfer_ragged_prefill_matches_cpu_reference() {
         // 2 ragged requests (6 + 10 tokens), GQA 4 qo / 2 kv heads, hd=256, causal.
         const HD: usize = 256;
@@ -403,7 +404,7 @@ mod tests {
                         dot += g * r;
                         na += g * g;
                         nb += r * r;
-                        max_rel = max_rel.max((g - r).abs() / (r.abs() + 1e-3) as f64);
+                        max_rel = max_rel.max((g - r).abs() / (r.abs() + 1e-3));
                     }
                     let cos = dot / (na.sqrt() * nb.sqrt() + 1e-12);
                     worst_cos = worst_cos.min(cos);
